+++ messaging-core/src/main/java/com/nabla/sdk/core/data/exception/CoroutineHelpers.kt
@@ -4,10 +4,14 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 
 internal fun <T> Result<T>.mapFailureAsNablaException(exceptionMapper: NablaExceptionMapper): Result<T> {
-    return when (val exception = exceptionOrNull()) {
-        null -> this
-        else -> Result.failure(exceptionMapper.map(exception))
-    }
+    return mapFailure(exceptionMapper::map)
+}
+
+internal fun <T> Result<T>.mapFailure(exceptionMapper: (Throwable) -> Throwable): Result<T> {
+    return fold(
+        onFailure = { Result.failure<T>(exceptionMapper(it)) },
+        onSuccess = { Result.success<T>(it) }
+    )
 }
 
 internal fun <T> Flow<T>.catchAndRethrowAsNablaException(exceptionMapper: NablaExceptionMapper) = catch {