+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/NablaMessagingClient.kt
@@ -1,6 +1,7 @@
 package com.nabla.sdk.messaging.core
 
 import androidx.annotation.CheckResult
+import com.benasher44.uuid.Uuid
 import com.nabla.sdk.core.NablaClient
 import com.nabla.sdk.core.domain.boundary.Logger
 import com.nabla.sdk.core.domain.entity.NablaException
@@ -48,10 +49,19 @@ public interface NablaMessagingClient {
      *
      * Reference the returned [Conversation.id] for further actions on that freshly created conversation: send message, mark as read, etc.
      *
+     * @param title optional - title for the conversation
+     * @param providerIdToAssign optional - [com.nabla.sdk.core.domain.entity.Provider.id] of a provider
+     *                           to which the conversation should be assigned.
+     *                           Make sure the specified provider has enough rights to be assigned a conversation.
+     *                           See [Roles and Permissions](https://docs.nabla.com/docs/roles-and-permissions).
+     *
      * @return [Result] of the operation, eventual errors will be of type [NablaException].
      */
     @CheckResult
-    public suspend fun createConversation(): Result<Conversation>
+    public suspend fun createConversation(
+        title: String? = null,
+        providerIdToAssign: Uuid? = null,
+    ): Result<Conversation>
 
     /**
      * Watch a conversation details.